+++ end_to_end_test/lib/enums.g.dart
@@ -157,6 +157,26 @@ final BuiltSet<FallbackEnum> _$fbValues =
   _$fbNo,
 ]);
 
+const DeprecatedEnum _$dYes = const DeprecatedEnum._('yes');
+const DeprecatedEnum _$dNo = const DeprecatedEnum._('no');
+
+DeprecatedEnum _$dValueOf(String name) {
+  switch (name) {
+    case 'yes':
+      return _$dYes;
+    case 'no':
+      return _$dNo;
+    default:
+      throw new ArgumentError(name);
+  }
+}
+
+final BuiltSet<DeprecatedEnum> _$dValues =
+    new BuiltSet<DeprecatedEnum>(const <DeprecatedEnum>[
+  _$dYes,
+  _$dNo,
+]);
+
 Serializer<TestEnum> _$testEnumSerializer = new _$TestEnumSerializer();
 Serializer<WireNameEnum> _$wireNameEnumSerializer =
     new _$WireNameEnumSerializer();
@@ -286,4 +306,4 @@ class _$FallbackEnumSerializer implements PrimitiveSerializer<FallbackEnum> {
       FallbackEnum.valueOf(serialized as String);
 }
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
+// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new